// packages/nx-plugin/src/generators/create-package/schema.d.ts
import type { Linter } from '@nrwl/linter';

export interface CreatePackageSchema {
  name: string;
  project: string;

  // options to create cli package, passed to js library generator
  directory?: string;
  skipFormat: boolean;
  tags?: string;
  unitTestRunner: 'jest' | 'none';
  linter: Linter;
  compiler: 'swc' | 'tsc';

  // options to create e2e project, passed to e2e project generator
  e2eTestRunner?: 'jest' | 'none';
}